Fix error with empty style definitions json file
An artifact in our own codebase could sometimes cause empty style
definitions json files. This will fail compilation of ex_css_modules
because decoding an empty string is not working.

To capture this edge case, let's check if the file is empty before
trying to decode OR generate the json file through a task.